Create new MathTraits header to provide compile-time information about Math types
